Cart page and profile page hw
