+++ Lesson32web/src/main/java/eshop/model/Cart.java
@@ -1,0 +1,18 @@
+package eshop.model;
+
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
+
+public class Cart implements Serializable {
+
+    private List<Product> products = new ArrayList<>();
+
+    public void addProduct(Product myProduct) {
+        products.add(myProduct);
+    }
+
+    public void deleteProduct(Product myProduct) {
+        products.remove(myProduct);
+    }
+}
