+++ Lesson32web/src/main/java/eshop/model/Product.java
@@ -1,4 +1,4 @@
-package loginform.model;
+package eshop.model;
 
 import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
@@ -13,9 +13,9 @@ import lombok.ToString;
 public class Product {
 
     private int id;
+    private String imageName;
     private String name;
     private String description;
     private BigDecimal price;
-    private String imageName;
-    private String categoryName;
+    private int categoryId;
 }
