+++ Lesson32web/src/main/webapp/cart.jsp
@@ -5,6 +5,8 @@
   Time: 0:45
   To change this template use File | Settings | File Templates.
 --%>
+<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
+<%@ taglib uri="http://java.sun.com/jsp/jstl/fmt" prefix="fmt" %>
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
 <html>
 <head>
@@ -17,6 +19,22 @@
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
 </head>
 <body>
+<nav class="navbar navbar-expand-lg navbar-light bg-light">
+    <div class="container-fluid">
+        <a class="navbar-brand" href="#">Navbar</a>
+        <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNavAltMarkup"
+                aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
+            <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
+            <div class="navbar-nav">
+                <a class="nav-link active" aria-current="page" href="home.jsp">Home</a>
+                <a class="nav-link" href="cart.jsp">Cart</a>
+                <a class="nav-link" href="cabinet.jsp">Profile</a>
+            </div>
+        </div>
+    </div>
+</nav>
 <section class="h-100" style="background-color: #eee;">
     <div class="container h-100 py-5">
         <div class="row d-flex justify-content-center align-items-center h-100">
@@ -29,15 +47,11 @@
                                 class="fas fa-angle-down mt-1"></i></a></p>
                     </div>
                 </div>
-
                 <div class="card rounded-3 mb-4">
-                    <c:forEach items="${sessionScope.productsCart}" var="product">
+                    <c:forEach items="${sessionScope.products}" var="product">
                         <div class="card-body p-4">
                             <div class="row d-flex justify-content-between align-items-center">
                                 <div class="col-md-2 col-lg-2 col-xl-2">
-                                        <%--                                    <img--%>
-                                        <%--                                            src="https://mdbcdn.b-cdn.net/img/Photos/new-templates/bootstrap-shopping-carts/img1.webp"--%>
-                                        <%--                                            class="img-fluid rounded-3" alt="Cotton T-shirt">--%>
                                     <img class="card-img" style="width:150px;height:120px"
                                          src="${contextPath}/images/${product.getImageName()}" alt="Card image">
                                 </div>
@@ -80,13 +94,11 @@
                         <button type="button" class="btn btn-outline-warning btn-lg ms-3">Apply</button>
                     </div>
                 </div>
-
                 <div class="card">
                     <div class="card-body">
                         <button type="button" class="btn btn-warning btn-block btn-lg">Proceed to Pay</button>
                     </div>
                 </div>
-
             </div>
         </div>
     </div>
