+++ Lesson32web/src/main/webapp/styles/cabinet.css
@@ -1,0 +1,10 @@
+* {
+  margin: 0;
+  box-sizing: border-box;
+  padding: 0;
+  font-family: 'Barlow', sans-serif;
+}
+
+h1 {
+  text-align: center;
+}
